Expose saved_model via tf.saved_model.
This also exposes meta_Graph_pb2.TensorInfo as tf.TensorInfo.
Change: 144344131
