+++ tensorflow/examples/saved_model/saved_model_half_plus_two.py
@@ -12,16 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Exports an example linear regression inference graph.
+r"""Exports an example linear regression inference graph.
 
-Exports a TensorFlow graph to /tmp/saved_model/half_plus_two/ based on the
-SavedModel format.
+Exports a TensorFlow graph to `/tmp/saved_model/half_plus_two/` based on the
+`SavedModel` format.
 
 This graph calculates,
+
+\\(
   y = a*x + b
+\\)
+
 and/or, independently,
+
+\\(
   y2 = a*x2 + c
-where a, b and c are variables with a=0.5 and b=2 and c=3.
+\\)
+
+where `a`, `b` and `c` are variables with `a=0.5` and `b=2` and `c=3`.
 
 Output from this program is typically used to exercise SavedModel load and
 execution code.
@@ -37,13 +45,6 @@ import sys
 
 import tensorflow as tf
 
-from tensorflow.core.protobuf import meta_graph_pb2
-from tensorflow.python.lib.io import file_io
-from tensorflow.python.saved_model import builder as saved_model_builder
-from tensorflow.python.saved_model import signature_constants
-from tensorflow.python.saved_model import signature_def_utils
-from tensorflow.python.saved_model import tag_constants
-from tensorflow.python.util import compat
 
 FLAGS = None
 
@@ -59,12 +60,12 @@ def _write_assets(assets_directory, assets_filename):
   Returns:
     The path to which the assets file was written.
   """
-  if not file_io.file_exists(assets_directory):
-    file_io.recursive_create_dir(assets_directory)
+  if not tf.python_io.file_exists(assets_directory):
+    tf.python_io.recursive_create_dir(assets_directory)
 
   path = os.path.join(
-      compat.as_bytes(assets_directory), compat.as_bytes(assets_filename))
-  file_io.write_string_to_file(path, "asset-file-contents")
+      tf.compat.as_bytes(assets_directory), tf.compat.as_bytes(assets_filename))
+  tf.python_io.write_string_to_file(path, "asset-file-contents")
   return path
 
 
@@ -75,7 +76,7 @@ def _generate_saved_model_for_half_plus_two(export_dir, as_text=False):
     export_dir: The directory to which the SavedModel should be written.
     as_text: Writes the SavedModel protocol buffer in text format to disk.
   """
-  builder = saved_model_builder.SavedModelBuilder(export_dir)
+  builder = tf.saved_model.builder.SavedModelBuilder(export_dir)
 
   with tf.Session(graph=tf.Graph()) as sess:
     # Set up the model parameters as variables to exercise variable loading
@@ -117,44 +118,48 @@ def _generate_saved_model_for_half_plus_two(export_dir, as_text=False):
 
     # Set up the signature for regression with input and output tensor
     # specification.
-    input_tensor = meta_graph_pb2.TensorInfo()
+    input_tensor = tf.TensorInfo()
     input_tensor.name = serialized_tf_example.name
-    signature_inputs = {signature_constants.REGRESS_INPUTS: input_tensor}
+    signature_inputs = {
+        tf.saved_model.signature_constants.REGRESS_INPUTS: input_tensor}
 
-    output_tensor = meta_graph_pb2.TensorInfo()
+    output_tensor = tf.TensorInfo()
     output_tensor.name = tf.identity(y).name
-    signature_outputs = {signature_constants.REGRESS_OUTPUTS: output_tensor}
-    signature_def = signature_def_utils.build_signature_def(
+    signature_outputs = {
+        tf.saved_model.signature_constants.REGRESS_OUTPUTS: output_tensor}
+    signature_def = tf.saved_model.signature_def_utils.build_signature_def(
         signature_inputs, signature_outputs,
-        signature_constants.REGRESS_METHOD_NAME)
+        tf.saved_model.signature_constants.REGRESS_METHOD_NAME)
 
     # Set up the signature for Predict with input and output tensor
     # specification.
-    predict_input_tensor = meta_graph_pb2.TensorInfo()
+    predict_input_tensor = tf.TensorInfo()
     predict_input_tensor.name = x.name
     predict_signature_inputs = {
         "x": predict_input_tensor
     }
 
-    predict_output_tensor = meta_graph_pb2.TensorInfo()
+    predict_output_tensor = tf.TensorInfo()
     predict_output_tensor.name = y.name
     predict_signature_outputs = {
         "y": predict_output_tensor
     }
-    predict_signature_def = signature_def_utils.build_signature_def(
-        predict_signature_inputs, predict_signature_outputs,
-        signature_constants.PREDICT_METHOD_NAME)
+    predict_signature_def = (
+        tf.saved_model.signature_def_utils.build_signature_def(
+            predict_signature_inputs, predict_signature_outputs,
+            tf.saved_model.signature_constants.PREDICT_METHOD_NAME))
 
     # Initialize all variables and then save the SavedModel.
     sess.run(tf.global_variables_initializer())
+    signature_def_map = {
+        tf.saved_model.signature_constants.REGRESS_METHOD_NAME:
+            signature_def,
+        tf.saved_model.signature_constants.DEFAULT_SERVING_SIGNATURE_DEF_KEY:
+            predict_signature_def
+    }
     builder.add_meta_graph_and_variables(
-        sess, [tag_constants.SERVING],
-        signature_def_map={
-            signature_constants.REGRESS_METHOD_NAME:
-                signature_def,
-            signature_constants.DEFAULT_SERVING_SIGNATURE_DEF_KEY:
-                predict_signature_def
-        },
+        sess, [tf.saved_model.tag_constants.SERVING],
+        signature_def_map=signature_def_map,
         assets_collection=tf.get_collection(tf.GraphKeys.ASSET_FILEPATHS),
         legacy_init_op=tf.group(assign_filename_op))
     builder.save(as_text)
